refactor: extract useRowLongPressMenu primitive from SearchList
